requirements.txt
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
pandas==2.2.3
pandas-gbq==0.26.0
tqdm==4.67.1
